+++ server/dataStructure/dict/DictDef.h
@@ -7,7 +7,7 @@
 
 typedef void (*scanProc)(void* priv, void* key, void* val);
 
-const int HASH_TABLE_INITIAL_SIZE = 4;      // hash table初始大小
-const int NEED_REHASH_RATIO = 5;            // rehash的触发比例
+const uint32_t HASH_TABLE_INITIAL_SIZE = 4;      // hash table初始大小
+const uint8_t NEED_REHASH_RATIO = 5;            // rehash的触发比例
 
 #endif //FLYDB_DICTDEF_H
